requirements.txt
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
numpy==1.21.6
pandas==1.3.5
pillow==9.5.0
scipy==1.7.3
scikit-learn==1.0.2
matplotlib==3.5.3